Allow backtrace filters to be configured
Currently all exception backtraces that match defined filters are
surpressed from the output. This means that in cases where some of
the filtered backtraces are actually required, we need to use the
`--backtrace` option. However, this significantly increases the
backtrace size and adds noise.

This change adds an option to use environment variable
`CUCUMBER_BACKTRACE_FILTERS` to configure which backtraces should be
filtered.